Add Mission Level Generator
Builds a mission graph and a space graph then maps the mission
onto the space. The space graph is then used to dig out tunnels
and dungeons in the level. The mission graph mapping is then
used to place the player, finish, keys, and locks into the space.

Right now this is very crude as a proof of concept. The mission
graph and space graph are hard coded and it doesn't do a good
job of placing the components in the space provided. It simply
places each mission component at the center of a dungeon. I just
wanted to get something like this working to start. Then we can
add more complexity to it.

+++ graph_to_level/spatial_graph_generator.py
@@ -34,7 +34,7 @@ class SpatialGraphGenerator:
 
     @staticmethod
     def align_nodes_to_grid(node_positions):
-        node_positions = np.round(node_positions)
+        node_positions = np.round(node_positions).astype(int)
         return node_positions
